# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nairobi Events Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listMyEvents
      tags: [events]
      summary: List events for authenticated planner
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/EventStatus"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Planner's events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event brief
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event details
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event brief (draft only)
      parameters:
        - $ref: "#/components/parameters/eventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventRequest"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}/submit:
    post:
      operationId: submitEventBrief
      tags: [events]
      summary: Submit event brief and trigger quote requests
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "200":
          description: Brief submitted, quote requests dispatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}/quotes:
    get:
      operationId: getEventQuotes
      tags: [events]
      summary: Get all quotes received for an event (comparison view)
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "200":
          description: Quotes grouped by category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventQuotesResponse"

  # ─── Quotes ─────────────────────────────────────────────────────────────────
...
  /bookings:
    get:
      operationId: listMyBookings
      tags: [bookings]
      summary: List bookings for authenticated user (planner or vendor)
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/BookingStatus"
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking details
      parameters:
        - $ref: "#/components/parameters/bookingId"
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          $ref: "#/components/responses/NotFound"
  /bookings/{bookingId}/confirm:
    post:
      operationId: confirmBooking
      tags: [bookings]
      summary: Confirm booking and initiate escrow payment
      parameters:
        - $ref: "#/components/parameters/bookingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmBookingRequest"
      responses:
        "200":
          description: Booking confirmed with payment intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingConfirmResponse"

  # ─── Reviews ─────────────────────────────────────────────────────────────────
...
  /admin/vendors/{vendorId}/approve:
    post:
      operationId: adminApproveVendor
      tags: [admin]
      summary: Approve a vendor
      parameters:
        - $ref: "#/components/parameters/vendorId"
      responses:
        "200":
          description: Vendor approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorProfile"
  /admin/vendors/{vendorId}/suspend:
    post:
      operationId: adminSuspendVendor
      tags: [admin]
      summary: Suspend a vendor
      parameters:
        - $ref: "#/components/parameters/vendorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required: [reason]
      responses:
        "200":
          description: Vendor suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorProfile"
...
components:
  parameters:
    vendorId:
      name: vendorId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    eventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    bookingId:
      name: bookingId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]

    # ─── Enums ────────────────────────────────────────────────────────────────
...
    EventType:
      type: string
      enum: [corporate, wedding, birthday, product_launch, conference, private_party, other]
    EventStatus:
      type: string
      enum: [draft, brief_submitted, quotes_requested, quotes_received, vendor_selected, booked, completed, cancelled]
    VendorCategory:
      type: string
      enum: [catering, mc, photography, videography, floristry, av_technical, tent_furniture, security, entertainment, decor, transportation, other]
    VendorStatus:
      type: string
      enum: [pending_review, approved, suspended, rejected]
    QuoteStatus:
      type: string
      enum: [requested, submitted, revision_requested, accepted, rejected, expired]
    BookingStatus:
      type: string
      enum: [pending, confirmed, in_escrow, completed, disputed, cancelled, refunded]

    # ─── User ─────────────────────────────────────────────────────────────────
...
    VendorProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        businessName:
          type: string
        category:
          $ref: "#/components/schemas/VendorCategory"
        description:
          type: string
        status:
          $ref: "#/components/schemas/VendorStatus"
        averageRating:
          type: string
          nullable: true
        totalReviews:
          type: integer
        totalBookings:
          type: integer
        isPremium:
          type: boolean
        city:
          type: string
        serviceAreas:
          type: array
          items:
            type: string
        websiteUrl:
          type: string
          nullable: true
        instagramHandle:
          type: string
          nullable: true
        portfolioUrls:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, businessName, category, description, status, totalReviews, totalBookings, isPremium, city, createdAt]
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        plannerId:
          type: string
          format: uuid
        title:
          type: string
        eventType:
          $ref: "#/components/schemas/EventType"
        status:
          $ref: "#/components/schemas/EventStatus"
        eventDate:
          type: string
          format: date
        eventEndDate:
          type: string
          format: date
          nullable: true
        venue:
          type: string
          nullable: true
        city:
          type: string
        guestCount:
          type: integer
        budgetMin:
          type: string
          nullable: true
        budgetMax:
          type: string
          nullable: true
        currency:
          type: string
        servicesNeeded:
          type: array
          items:
            type: string
        specialRequirements:
          type: string
          nullable: true
        isEmergency:
          type: boolean
        quotesDeadline:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, plannerId, title, eventType, status, eventDate, city, guestCount, currency, servicesNeeded, isEmergency, createdAt, updatedAt]
    CreateEventRequest:
      type: object
      properties:
        title:
          type: string
        eventType:
          $ref: "#/components/schemas/EventType"
        eventDate:
          type: string
          format: date
        eventEndDate:
          type: string
          format: date
        venue:
          type: string
        city:
          type: string
        guestCount:
          type: integer
        budgetMin:
          type: string
        budgetMax:
          type: string
        currency:
          type: string
          default: KES
        servicesNeeded:
          type: array
          items:
            type: string
        specialRequirements:
          type: string
        isEmergency:
          type: boolean
          default: false
      required: [title, eventType, eventDate, city, guestCount, servicesNeeded]
    UpdateEventRequest:
      type: object
      properties:
        title:
          type: string
        eventDate:
          type: string
          format: date
        venue:
          type: string
        guestCount:
          type: integer
        budgetMin:
          type: string
        budgetMax:
          type: string
        servicesNeeded:
          type: array
          items:
            type: string
        specialRequirements:
          type: string
    EventListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [events, total, page, limit]
    EventQuotesResponse:
      type: object
      properties:
        eventId:
          type: string
          format: uuid
        quotesByCategory:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/Quote"
      required: [eventId, quotesByCategory]

    # ─── Quote ────────────────────────────────────────────────────────────────
    QuoteLineItem:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
      required: [description, quantity, unitPrice, total]
    Quote:
      type: object
      properties:
        id:
          type: string
          format: uuid
        quoteRequestId:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        vendorId:
          type: string
          format: uuid
        category:
          $ref: "#/components/schemas/VendorCategory"
        status:
          $ref: "#/components/schemas/QuoteStatus"
        totalAmount:
          type: string
        currency:
          type: string
        depositPercent:
          type: integer
          nullable: true
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/QuoteLineItem"
        inclusions:
          type: array
          items:
            type: string
        exclusions:
          type: array
          items:
            type: string
        terms:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        version:
          type: integer
        submittedAt:
          type: string
          format: date-time
        vendor:
          $ref: "#/components/schemas/VendorProfile"
      required: [id, quoteRequestId, eventId, vendorId, category, status, totalAmount, currency, lineItems, version, submittedAt]
...
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        quoteId:
          type: string
          format: uuid
        vendorId:
          type: string
          format: uuid
        plannerId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/BookingStatus"
        totalAmount:
          type: string
        platformFeeAmount:
          type: string
        vendorPayoutAmount:
          type: string
        currency:
          type: string
        stripePaymentIntentId:
          type: string
          nullable: true
        contractUrl:
          type: string
          nullable: true
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, eventId, quoteId, vendorId, plannerId, status, totalAmount, platformFeeAmount, vendorPayoutAmount, currency, createdAt]
    ConfirmBookingRequest:
      type: object
      properties:
        paymentMethodId:
          type: string
      required: [paymentMethodId]
    BookingConfirmResponse:
      type: object
      properties:
        booking:
          $ref: "#/components/schemas/Booking"
        clientSecret:
          type: string
      required: [booking, clientSecret]

    # ─── Review ───────────────────────────────────────────────────────────────
...