# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nairobi Events Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: string;
  eventId: string;
  quoteId: string;
  vendorId: string;
  plannerId: string;
  status: BookingStatus;
  totalAmount: string;
  platformFeeAmount: string;
  vendorPayoutAmount: string;
  currency: string;
  stripePaymentIntentId?: string | null;
  contractUrl?: string | null;
  confirmedAt?: Date | null;
  completedAt?: Date | null;
  createdAt: Date;
}