# lib/api-zod/src/generated/types/bookingStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nairobi Events Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  in_escrow: "in_escrow",
  completed: "completed",
  disputed: "disputed",
  cancelled: "cancelled",
  refunded: "refunded",
} as const;