# lib/api-zod/src/generated/types/eventQuotesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nairobi Events Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { EventQuotesResponseQuotesByCategory } from "./eventQuotesResponseQuotesByCategory";

export interface EventQuotesResponse {
  eventId: string;
  quotesByCategory: EventQuotesResponseQuotesByCategory;
}