# lib/api-zod/src/generated/types/vendorProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nairobi Events Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { VendorCategory } from "./vendorCategory";
import type { VendorStatus } from "./vendorStatus";

export interface VendorProfile {
  id: string;
  userId: string;
  businessName: string;
  category: VendorCategory;
  description: string;
  status: VendorStatus;
  averageRating?: string | null;
  totalReviews: number;
  totalBookings: number;
  isPremium: boolean;
  city: string;
  serviceAreas?: string[];
  websiteUrl?: string | null;
  instagramHandle?: string | null;
  portfolioUrls?: string[];
  createdAt: Date;
}